Restored $rand_key functionality
It is only used fro the spam trap input but had been randomly generated
each call; and therefore would never work. Now it is dynamically set in
membersite_config BUT uses the year and day of year as its seed. In this
way it is not entirely static; but should be between form loads and
submissions.